fix asset deployment
I didn't want to clutter the root of the repository with the metric
****-ton of files that's needed for proper favicon deployment, but so it
happens that I can't track them in _site and specify that they should be
cached build artifacts in _config.yml. (That's a lie, actually: it works
*locally* but *of course* it won't necessarily work on deployment if it
works locally. More specifically it doesn't work when I deploy to GH
pages and I don't want to spend time bouncing around caching issues when
I'm reasonably hopeful that this will actually work. Hopefully...)