css: fix rendering on mobile Safari
It turns out that when only "Roboto" is imported, browsers have to use a
"faux style" for the bold version of Roboto, by applying transforms to
the stroke width and text offset (I'm not sure if those are the correct
technical terms, but so be it). This resulted in Roboto headings looking
really weird in mobile Safari.

The fix is to actually download the font for all font weights/styles I
intend to use, i.e. actually download the bold version of Roboto.